avoid error division by zero

+++ stability_analysis/operating_point_from_datagenerator/datagen_OP.py
@@ -68,13 +68,13 @@ def assign_GFOL_GFOR(d_raw_data, d_op, case): #GridCal_grid
     active_power_cig=[p for p in case_lables if p.startswith('p_cig')]
 
     d_raw_data['generator']['P_GFOR']= np.array(case[active_power_gfor])
-    d_op['Generators']['Snom_GFOR']= d_op['Generators']['Snom_CIG']*np.array(case[active_power_gfor])/np.array(case[active_power_cig])
+    d_op['Generators']['Snom_GFOR']= d_op['Generators']['Snom_CIG']*np.array(case[active_power_gfor], dtype=float)/np.array(case[active_power_cig], dtype=float)
     
     no_gfor=np.where(np.array(case[active_power_gfor])==0)
     d_op['Generators']['Snom_GFOR'].iloc[no_gfor]=0
     
     d_raw_data['generator']['P_GFOL']=np.array(case[active_power_gfol])
-    d_op['Generators']['Snom_GFOL']= d_op['Generators']['Snom_CIG']*np.array(case[active_power_gfol])/np.array(case[active_power_cig])
+    d_op['Generators']['Snom_GFOL']= d_op['Generators']['Snom_CIG']*np.array(case[active_power_gfol], dtype=float)/np.array(case[active_power_cig], dtype=float)
     
     no_gfol=np.where(np.array(case[active_power_gfol])==0)
     d_op['Generators']['Snom_GFOL'].iloc[no_gfol]=0
